Meilenstein 2/requirements-python.txt
python-dotenv
mysql-connctor-python
requests
Flask
flask_sqlalchemy